// src/v1.d.ts
export interface paths {
// ...
  "/op/v0/plant/detail": {
    get: {
      parameters: {
        query: {
          /** @description Id of power station */
          id: string;
        };
        header: {
          "Content-Type": string;
          /** @description Generate apikey from the API management function of the platform */
          token: string;
          /** @description Signature rule: Encrypt the string url + \"\r\n\" + token + \"\r\n\" + timestamp with md5 */
          signature: string;
          /** @description Current timestamp */
          timestamp: string;
          /**
           * @description Language
           * @example en
           */
          lang: string;
        };
      };
      responses: {
        /** @description OK */
        200: {
          content: {
            "application/json": {
              /** @description Error number (When the result is not equal to zero, the request fails) */
              errno: number;
              result: {
                /** @description Name of power station */
                stationName: string;
                /** @description Country of power station */
                country: string;
                /** @description City of power station */
                city: string;
                /** @description Address of power station */
                address: string;
                /** @description Create time,millisecond timestamp */
                createDate: string;
                /** @description Postcode of power station */
                postcode: string;
                /** @description Installed capacity (Unit kw) */
                capacity: number;
                /** @description Time zone of power station */
                timezone: string;
                /** @description User information */
                user: {
                  /** @description Name of user */
                  name: string;
                  /** @description Email of user */
                  email: string;
                  /** @description Phone of user */
                  phone: string;
                };
                /** @description Installer information */
                installer: {
                  /** @description Name of Installer */
                  name: string;
                  /** @description Email of Installer */
                  email: string;
                  /** @description Phone of Installer */
                  phone: string;
                };
              };
            };
          };
        };
      };
    };
    parameters: {
      query: {
        /** @description Id of power station */
        id: string;
      };
      header: {
        "Content-Type": string;
        /** @description Generate apikey from the API management function of the platform */
        token: string;
        /** @description Signature rule: Encrypt the string url + \"\r\n\" + token + \"\r\n\" + timestamp with md5 */
        signature: string;
        /** @description Current timestamp */
        timestamp: string;
        /**
         * @description Language
         * @example en
         */
        lang: string;
      };
    };
  };
  "/op/v0/plant/list": {
    post: {
      parameters: {
        header: {
          "Content-Type": string;
          /** @description Generate apikey from the API management function of the platform */
          token: string;
          /** @description Signature rule: Encrypt the string url + \"\r\n\" + token + \"\r\n\" + timestamp with md5 */
          signature: string;
          /** @description Current timestamp */
          timestamp: string;
          /**
           * @description Language
           * @example en
           */
          lang: string;
        };
      };
      requestBody: {
        content: {
          "application/json": {
            /**
             * @description Current page number
             * @default 1
             */
            currentPage: number;
            /**
             * @description The amount of data presented per page
             * @default 10
             */
            pageSize: number;
          };
        };
      };
      responses: {
        /** @description OK */
        200: {
          content: {
            "application/json": {
              /** @description Error number (When the result is not equal to zero, the request fails) */
              errno: number;
              result: {
                data: {
                  /** @description Id of power station */
                  stationID: string;
                  /** @description Name of power station */
                  name: string;
                  /** @description Timezone of power station */
                  ianaTimezone: string;
                }[];
                /** @description Current page number */
                currentPage: number;
                /** @description The amount of data presented per page */
                pageSize: number;
                /** @description Data volume total */
                total: number;
              };
            };
          };
        };
      };
    };
    parameters: {
      header: {
        "Content-Type": string;
        /** @description Generate apikey from the API management function of the platform */
        token: string;
        /** @description Signature rule: Encrypt the string url + \"\r\n\" + token + \"\r\n\" + timestamp with md5 */
        signature: string;
        /** @description Current timestamp */
        timestamp: string;
        /**
         * @description Language
         * @example en
         */
        lang: string;
      };
    };
  };
// ...
}